Fix another oopsie

+++ rsna-2024/training_utils.py
@@ -42,7 +42,7 @@ def model_validation_loss(model, val_loader, loss_fns):
 
     for images, label in val_loader:
         # !TODO: Do this in the data loader
-        label = label.type(torch.FloatTensor).to(device)
+        label = label.type(torch.LongTensor).to(device)
 
         output = model(images.to(device))
         loss = sum([loss_fn(output, label) for loss_fn in loss_fns])
